+++ tests/testthat/test-countryname.R
@@ -18,7 +18,7 @@ test_that('input: character ', {
 
     x <- c('ジンバブエ', 'Afeganistãu',  'Barbadas', 'Sverige', 'UK',
            'il-Georgia tan-Nofsinhar u l-Gżejjer Sandwich tan-Nofsinhar')
-    answers <- c("Zimbabwe", "Afghanistan", "Barbados", "Sweden", "UK", 
+    answers <- c("Zimbabwe", "Afghanistan", "Barbados", "Sweden", "United Kingdom", 
                  "South Georgia & South Sandwich Islands")
     expect_identical(countryname(x), answers)
 
@@ -26,11 +26,11 @@ test_that('input: character ', {
 
 test_that('input: factor vector', {
 
-    x <- c('ジンバブエ', 'Afeganistãu',  'Barbadas', 'Sverige', 'UK',
+    x <- c('ジンバブエ', 'Afeganistãu',  'Barbadas', 'Sverige', 'United Kingdom',
            'il-Georgia tan-Nofsinhar u l-Gżejjer Sandwich tan-Nofsinhar')
     x <- factor(x)
 
-    answers <- c("Zimbabwe", "Afghanistan", "Barbados", "Sweden", "UK", 
+    answers <- c("Zimbabwe", "Afghanistan", "Barbados", "Sweden", "United Kingdom", 
                  "South Georgia & South Sandwich Islands")
     expect_identical(countryname(x), answers)
 
@@ -43,8 +43,25 @@ test_that('input: tibble ', {
            'il-Georgia tan-Nofsinhar u l-Gżejjer Sandwich tan-Nofsinhar')
     x <- tibble(x)
 
-    answers <- c("Zimbabwe", "Afghanistan", "Barbados", "Sweden", "UK", 
+    answers <- c("Zimbabwe", "Afghanistan", "Barbados", "Sweden", "United Kingdom", 
                  "South Georgia & South Sandwich Islands")
     expect_identical(countryname(x$x), answers)
 
+})
+
+
+test_that("issue 336", {
+    a <- countrycode("antarctica", origin = "country.name", destination = "cowc", warn = FALSE)
+    expect_true(is.na(a))
+
+    a <- countryname("antarctica", destination = "cowc", warn = FALSE)
+    expect_true(is.na(a))
+
+    a <- countryname("xyz", destination = "cowc", warn = FALSE)
+    expect_true(is.na(a))
+
+    x <- c("canada", "antarctica")
+    expect_identical(countryname(x), c("Canada", "Antarctica"))
+    expect_identical(countryname(x, destination = "cowc", warn = FALSE), c("CAN", NA))
+    expect_identical(countryname(x, destination = "cowc", warn = FALSE, nomatch = x), c("CAN", "antarctica"))
 })